+++ cmd/manager/main.go
@@ -133,7 +133,7 @@ func main() {
 		Log:              logger.Named("controllers").Named("AtlasDeployment").Sugar(),
 		Scheme:           mgr.GetScheme(),
 		AtlasDomain:      config.AtlasDomain,
-		Revision:         config.Revision,
+		OperatorClass:    config.OperatorClass,
 		GlobalAPISecret:  config.GlobalAPISecret,
 		ResourceWatcher:  watch.NewResourceWatcher(),
 		GlobalPredicates: globalPredicates,
@@ -148,7 +148,7 @@ func main() {
 		Log:              logger.Named("controllers").Named("AtlasProject").Sugar(),
 		Scheme:           mgr.GetScheme(),
 		AtlasDomain:      config.AtlasDomain,
-		Revision:         config.Revision,
+		OperatorClass:    config.OperatorClass,
 		ResourceWatcher:  watch.NewResourceWatcher(),
 		GlobalAPISecret:  config.GlobalAPISecret,
 		GlobalPredicates: globalPredicates,
@@ -163,7 +163,7 @@ func main() {
 		Log:              logger.Named("controllers").Named("AtlasDatabaseUser").Sugar(),
 		Scheme:           mgr.GetScheme(),
 		AtlasDomain:      config.AtlasDomain,
-		Revision:         config.Revision,
+		OperatorClass:    config.OperatorClass,
 		ResourceWatcher:  watch.NewResourceWatcher(),
 		GlobalAPISecret:  config.GlobalAPISecret,
 		GlobalPredicates: globalPredicates,
@@ -192,7 +192,7 @@ func main() {
 
 type Config struct {
 	AtlasDomain          string
-	Revision             string
+	OperatorClass        string
 	EnableLeaderElection bool
 	MetricsAddr          string
 	Namespace            string
@@ -208,7 +208,7 @@ func parseConfiguration() Config {
 	var globalAPISecretName string
 	config := Config{}
 	flag.StringVar(&config.AtlasDomain, "atlas-domain", "https://cloud.mongodb.com/", "the Atlas URL domain name (with slash in the end).")
-	flag.StringVar(&config.Revision, "revision", "", "The revision, this can be used to only reconcile on resources with mongodb.com/atlas-revision=<revision>")
+	flag.StringVar(&config.OperatorClass, "operator-class", "", "The operator class, this can be used to only reconcile on resources with mongodb.com/atlas-operator-class=<operator-class>")
 	flag.StringVar(&config.MetricsAddr, "metrics-bind-address", ":8080", "The address the metric endpoint binds to.")
 	flag.StringVar(&config.ProbeAddr, "health-probe-bind-address", ":8081", "The address the probe endpoint binds to.")
 	flag.StringVar(&globalAPISecretName, "global-api-secret-name", "", "The name of the Secret that contains Atlas API keys. "+