+++ pkg/controller/customresource/customresource.go
@@ -24,7 +24,7 @@ const (
 	ReconciliationPolicyAnnotation = "mongodb.com/atlas-reconciliation-policy"
 	ResourceVersion                = "app.kubernetes.io/version"
 	ResourceVersionOverride        = "mongodb.com/atlas-resource-version-policy"
-	RevisionAnnotation             = "mongodb.com/atlas-revision"
+	RevisionAnnotation             = "mongodb.com/atlas-operator-class"
 
 	ResourcePolicyKeep       = "keep"
 	ReconciliationPolicySkip = "skip"
@@ -129,9 +129,9 @@ func ResourceVersionIsValid(resource mdbv1.AtlasCustomResource) (bool, error) {
 }
 
 // ReconciliationShouldBeSkipped returns 'true' if reconciliation should be skipped for this resource.
-func ReconciliationShouldBeSkipped(resource mdbv1.AtlasCustomResource, revision string) bool {
+func ReconciliationShouldBeSkipped(resource mdbv1.AtlasCustomResource, operatorClass string) bool {
 	if v, ok := resource.GetAnnotations()[RevisionAnnotation]; ok {
-		if v != revision {
+		if v != operatorClass {
 			return true
 		}
 	}