+++ docs/release-notes/release-notes.md
@@ -4,7 +4,7 @@
 
 * Fix connection secret creation (#774)
 * Fix mininum version of Openshift (#761)
-* add revision to only reconcile resources with mongodb.com/atlas-revision=<revision> annotation (#842)
+* add operator class to only reconcile resources with mongodb.com/atlas-operator-class=<operator-class> annotation (#842)
 
 ## AtlasProject Resource